+++ modules/imgproc/src/pyramids.cpp
@@ -442,7 +442,7 @@ static bool ocl_pyrDown( InputArray _src, OutputArray _dst, const Size& _dsz, in
     k.args(ocl::KernelArg::ReadOnly(src), ocl::KernelArg::WriteOnly(dst));
 
     size_t localThreads[2]  = { local_size, 1 };
-    size_t globalThreads[2] = { src.cols, dst.rows };
+    size_t globalThreads[2] = { static_cast<size_t>(src.cols), static_cast<size_t>(dst.rows) };
     return k.run(2, globalThreads, localThreads, false);
 }
