+++ modules/imgproc/src/templmatch.cpp
@@ -107,7 +107,7 @@ static bool matchTemplateNaive_CCORR(InputArray _image, InputArray _templ, Outpu
     k.args(ocl::KernelArg::ReadOnlyNoSize(image), ocl::KernelArg::ReadOnly(templ),
            ocl::KernelArg::WriteOnly(result));
 
-    size_t globalsize[2] = { result.cols, result.rows };
+    size_t globalsize[2] = { static_cast<size_t>(result.cols), static_cast<size_t>(result.rows) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -136,7 +136,7 @@ static bool matchTemplate_CCORR_NORMED(InputArray _image, InputArray _templ, Out
     k.args(ocl::KernelArg::ReadOnlyNoSize(image_sqsums), ocl::KernelArg::ReadWrite(result),
            templ.rows, templ.cols, ocl::KernelArg::PtrReadOnly(templ_sqsum));
 
-    size_t globalsize[2] = { result.cols, result.rows };
+    size_t globalsize[2] = { static_cast<size_t>(result.cols), static_cast<size_t>(result.rows) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -161,7 +161,7 @@ static bool matchTemplateNaive_SQDIFF(InputArray _image, InputArray _templ, Outp
     k.args(ocl::KernelArg::ReadOnlyNoSize(image), ocl::KernelArg::ReadOnly(templ),
            ocl::KernelArg::WriteOnly(result));
 
-    size_t globalsize[2] = { result.cols, result.rows };
+    size_t globalsize[2] = { static_cast<size_t>(result.cols), static_cast<size_t>(result.rows) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -190,7 +190,7 @@ static bool matchTemplate_SQDIFF_NORMED(InputArray _image, InputArray _templ, Ou
     k.args(ocl::KernelArg::ReadOnlyNoSize(image_sqsums), ocl::KernelArg::ReadWrite(result),
            templ.rows, templ.cols, ocl::KernelArg::PtrReadOnly(templ_sqsum));
 
-    size_t globalsize[2] = { result.cols, result.rows };
+    size_t globalsize[2] = { static_cast<size_t>(result.cols), static_cast<size_t>(result.rows) };
 
     return k.run(2, globalsize, NULL, false);
 }
@@ -244,7 +244,7 @@ static bool matchTemplate_CCOEFF(InputArray _image, InputArray _templ, OutputArr
                    templ_sum[0], templ_sum[1], templ_sum[2], templ_sum[3]);
     }
 
-    size_t globalsize[2] = { result.cols, result.rows };
+    size_t globalsize[2] = { static_cast<size_t>(result.cols), static_cast<size_t>(result.rows) };
     return k.run(2, globalsize, NULL, false);
 }
 
@@ -322,7 +322,7 @@ static bool matchTemplate_CCOEFF_NORMED(InputArray _image, InputArray _templ, Ou
                    templ_sum[0], templ_sum[1], templ_sum[2], templ_sum[3], templ_sqsum_sum);
     }
 
-    size_t globalsize[2] = { result.cols, result.rows };
+    size_t globalsize[2] = { static_cast<size_t>(result.cols), static_cast<size_t>(result.rows) };
     return k.run(2, globalsize, NULL, false);
 }
